chore: simplify builder setup to match node (part 1)
