+++ bin/builder/src/cli.rs
@@ -1,18 +1,81 @@
-//! Additional Node command arguments for the OP builder.
-//!
-//! Copied from `OptimismNode` to allow easy extension.
+//! Contains the CLI arguments
 
 use core::{convert::TryFrom, net::SocketAddr, time::Duration};
 
 use base_builder_core::{BuilderConfig, FlashblocksConfig, TxDataStore};
 use reth_optimism_node::args::RollupArgs;
 
-use crate::{FlashblocksArgs, TelemetryArgs};
+/// Parameters for Flashblocks configuration.
+///
+/// The names in the struct are prefixed with `flashblocks` to avoid conflicts
+/// with the legacy standard builder configuration (now removed) since these args are
+/// flattened into the main `Args` struct with the other rollup/node args.
+#[derive(Debug, Clone, PartialEq, Eq, clap::Args)]
+pub struct FlashblocksArgs {
+    /// Flashblocks is always enabled; these options tune its behavior.
+    /// The port that we bind to for the websocket server that provides flashblocks
+    #[arg(long = "flashblocks.port", env = "FLASHBLOCKS_WS_PORT", default_value = "1111")]
+    pub flashblocks_port: u16,
+
+    /// The address that we bind to for the websocket server that provides flashblocks
+    #[arg(long = "flashblocks.addr", env = "FLASHBLOCKS_WS_ADDR", default_value = "127.0.0.1")]
+    pub flashblocks_addr: String,
+
+    /// flashblock block time in milliseconds
+    #[arg(long = "flashblocks.block-time", default_value = "250", env = "FLASHBLOCK_BLOCK_TIME")]
+    pub flashblocks_block_time: u64,
+
+    /// Builder would always try to produce fixed number of flashblocks without regard to time of
+    /// FCU arrival.
+    /// In cases of late FCU it could lead to partially filled blocks.
+    #[arg(long = "flashblocks.fixed", default_value = "false", env = "FLASHBLOCK_FIXED")]
+    pub flashblocks_fixed: bool,
+
+    /// Time by which blocks would be completed earlier in milliseconds.
+    ///
+    /// This time is used to account for latencies and would be deducted from total block
+    /// building time before calculating number of fbs.
+    #[arg(long = "flashblocks.leeway-time", default_value = "75", env = "FLASHBLOCK_LEEWAY_TIME")]
+    pub flashblocks_leeway_time: u64,
+
+    /// Whether to disable state root calculation for each flashblock
+    #[arg(
+        long = "flashblocks.disable-state-root",
+        default_value = "false",
+        env = "FLASHBLOCKS_DISABLE_STATE_ROOT"
+    )]
+    pub flashblocks_disable_state_root: bool,
+
+    /// Whether to compute state root only when `get_payload` is called (finalization).
+    /// When enabled, flashblocks are built without state root, but the final payload
+    /// returned by `get_payload` will have the state root computed.
+    /// Requires --flashblocks.disable-state-root to be effective.
+    #[arg(
+        long = "flashblocks.compute-state-root-on-finalize",
+        default_value = "false",
+        env = "FLASHBLOCKS_COMPUTE_STATE_ROOT_ON_FINALIZE"
+    )]
+    pub flashblocks_compute_state_root_on_finalize: bool,
+}
+
+impl Default for FlashblocksArgs {
+    fn default() -> Self {
+        Self {
+            flashblocks_port: 1111,
+            flashblocks_addr: "127.0.0.1".to_string(),
+            flashblocks_block_time: 250,
+            flashblocks_fixed: false,
+            flashblocks_leeway_time: 75,
+            flashblocks_disable_state_root: false,
+            flashblocks_compute_state_root_on_finalize: false,
+        }
+    }
+}
 
 /// Parameters for rollup configuration
 #[derive(Debug, Clone, PartialEq, Eq, clap::Args)]
 #[command(next_help_heading = "Rollup")]
-pub struct BuilderArgs {
+pub struct Args {
     /// Rollup configuration
     #[command(flatten)]
     pub rollup_args: RollupArgs,
@@ -37,15 +100,16 @@ pub struct BuilderArgs {
     #[arg(long = "builder.tx-data-store-buffer-size", default_value = "10000")]
     pub tx_data_store_buffer_size: usize,
 
+    /// Inverted sampling frequency in blocks. 1 - each block, 100 - every 100th block.
+    #[arg(long = "telemetry.sampling-ratio", env = "SAMPLING_RATIO", default_value = "100")]
+    pub sampling_ratio: u64,
+
     /// Flashblocks configuration
     #[command(flatten)]
     pub flashblocks: FlashblocksArgs,
-    /// Telemetry configuration
-    #[command(flatten)]
-    pub telemetry: TelemetryArgs,
 }
 
-impl Default for BuilderArgs {
+impl Default for Args {
     fn default() -> Self {
         Self {
             rollup_args: RollupArgs::default(),
@@ -54,23 +118,23 @@ impl Default for BuilderArgs {
             extra_block_deadline_secs: 20,
             enable_resource_metering: false,
             tx_data_store_buffer_size: 10000,
+            sampling_ratio: 100,
             flashblocks: FlashblocksArgs::default(),
-            telemetry: TelemetryArgs::default(),
         }
     }
 }
 
-impl TryFrom<BuilderArgs> for BuilderConfig {
+impl TryFrom<Args> for BuilderConfig {
     type Error = eyre::Report;
 
-    fn try_from(args: BuilderArgs) -> Result<Self, Self::Error> {
-        let flashblocks = FlashblocksConfig::try_from(args.clone())?;
+    fn try_from(args: Args) -> Result<Self, Self::Error> {
+        let flashblocks = FlashblocksConfig::try_from(&args)?;
         Ok(Self {
             block_time: Duration::from_millis(args.chain_block_time),
             block_time_leeway: Duration::from_secs(args.extra_block_deadline_secs),
             da_config: Default::default(),
             gas_limit_config: Default::default(),
-            sampling_ratio: args.telemetry.sampling_ratio,
+            sampling_ratio: args.sampling_ratio,
             max_gas_per_txn: args.max_gas_per_txn,
             tx_data_store: TxDataStore::new(
                 args.enable_resource_metering,
@@ -81,10 +145,10 @@ impl TryFrom<BuilderArgs> for BuilderConfig {
     }
 }
 
-impl TryFrom<BuilderArgs> for FlashblocksConfig {
+impl TryFrom<&Args> for FlashblocksConfig {
     type Error = eyre::Report;
 
-    fn try_from(args: BuilderArgs) -> Result<Self, Self::Error> {
+    fn try_from(args: &Args) -> Result<Self, Self::Error> {
         let interval = Duration::from_millis(args.flashblocks.flashblocks_block_time);
 
         let ws_addr = SocketAddr::new(
@@ -94,20 +158,15 @@ impl TryFrom<BuilderArgs> for FlashblocksConfig {
 
         let leeway_time = Duration::from_millis(args.flashblocks.flashblocks_leeway_time);
 
-        let fixed = args.flashblocks.flashblocks_fixed;
-
-        let disable_state_root = args.flashblocks.flashblocks_disable_state_root;
-
-        let compute_state_root_on_finalize =
-            args.flashblocks.flashblocks_compute_state_root_on_finalize;
-
         Ok(Self {
             ws_addr,
             interval,
             leeway_time,
-            fixed,
-            disable_state_root,
-            compute_state_root_on_finalize,
+            fixed: args.flashblocks.flashblocks_fixed,
+            disable_state_root: args.flashblocks.flashblocks_disable_state_root,
+            compute_state_root_on_finalize: args
+                .flashblocks
+                .flashblocks_compute_state_root_on_finalize,
         })
     }
 }
@@ -117,15 +176,14 @@ mod tests {
     use rstest::rstest;
 
     use super::*;
-    use crate::FlashblocksArgs;
 
-    fn convert(args: BuilderArgs) -> BuilderConfig {
+    fn convert(args: Args) -> BuilderConfig {
         BuilderConfig::try_from(args).expect("conversion should succeed")
     }
 
     #[test]
     fn default_args_produce_valid_config() {
-        let config = convert(BuilderArgs::default());
+        let config = convert(Args::default());
         assert_eq!(config.block_time, Duration::from_millis(1000));
         assert!(config.max_gas_per_txn.is_none());
     }
@@ -135,7 +193,7 @@ mod tests {
     #[case::block_time_2s(2000, 2000)]
     #[case::block_time_250ms(250, 250)]
     fn chain_block_time_maps_to_block_time(#[case] input_ms: u64, #[case] expected_ms: u64) {
-        let args = BuilderArgs { chain_block_time: input_ms, ..Default::default() };
+        let args = Args { chain_block_time: input_ms, ..Default::default() };
         let config = convert(args);
         assert_eq!(config.block_time, Duration::from_millis(expected_ms));
     }
@@ -145,7 +203,7 @@ mod tests {
     #[case::none(None, None)]
     #[case::large_gas(Some(1_000_000), Some(1_000_000))]
     fn max_gas_per_txn_maps_correctly(#[case] input: Option<u64>, #[case] expected: Option<u64>) {
-        let args = BuilderArgs { max_gas_per_txn: input, ..Default::default() };
+        let args = Args { max_gas_per_txn: input, ..Default::default() };
         let config = convert(args);
         assert_eq!(config.max_gas_per_txn, expected);
     }
@@ -155,7 +213,7 @@ mod tests {
     #[case::leeway_10s(10, 10)]
     #[case::leeway_0s(0, 0)]
     fn extra_block_deadline_maps_to_leeway(#[case] input_secs: u64, #[case] expected_secs: u64) {
-        let args = BuilderArgs { extra_block_deadline_secs: input_secs, ..Default::default() };
+        let args = Args { extra_block_deadline_secs: input_secs, ..Default::default() };
         let config = convert(args);
         assert_eq!(config.block_time_leeway, Duration::from_secs(expected_secs));
     }
@@ -165,7 +223,7 @@ mod tests {
     #[case::interval_200ms(200, 200)]
     #[case::interval_250ms(250, 250)]
     fn flashblocks_interval_maps_correctly(#[case] input_ms: u64, #[case] expected_ms: u64) {
-        let args = BuilderArgs {
+        let args = Args {
             flashblocks: FlashblocksArgs { flashblocks_block_time: input_ms, ..Default::default() },
             ..Default::default()
         };
@@ -177,7 +235,7 @@ mod tests {
     #[case::fixed_true(true, true)]
     #[case::fixed_false(false, false)]
     fn flashblocks_fixed_mode_maps_correctly(#[case] input: bool, #[case] expected: bool) {
-        let args = BuilderArgs {
+        let args = Args {
             flashblocks: FlashblocksArgs { flashblocks_fixed: input, ..Default::default() },
             ..Default::default()
         };
@@ -196,7 +254,7 @@ mod tests {
         #[case] disable_expected: bool,
         #[case] finalize_expected: bool,
     ) {
-        let args = BuilderArgs {
+        let args = Args {
             flashblocks: FlashblocksArgs {
                 flashblocks_disable_state_root: disable_input,
                 flashblocks_compute_state_root_on_finalize: finalize_input,
@@ -211,7 +269,7 @@ mod tests {
 
     #[test]
     fn combined_overrides_work_together() {
-        let args = BuilderArgs {
+        let args = Args {
             chain_block_time: 2000,
             max_gas_per_txn: Some(100000),
             extra_block_deadline_secs: 10,
