--- crates/builder/cli/Cargo.toml
@@ -1,51 +1,0 @@
-[package]
-name = "base-builder-cli"
-description = "CLI argument definitions for the Base builder"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[lints]
-workspace = true
-
-[dependencies]
-# Workspace
-base-cli-utils.workspace = true
-base-builder-core.workspace = true
-
-# CLI
-clap.workspace = true
-
-# reth
-reth-tracing-otlp.workspace = true
-reth-cli-commands.workspace = true
-reth-node-builder.workspace = true
-reth-db = { workspace = true, features = ["op"] }
-
-# optimism reth
-reth-optimism-cli = { workspace = true, features = ["otlp"] }
-reth-optimism-node.workspace = true
-reth-optimism-chainspec.workspace = true
-reth-optimism-rpc = { workspace = true, features = ["client"] }
-reth-optimism-txpool.workspace = true
-
-# tracing
-tracing.workspace = true
-tracing-subscriber.workspace = true
-
-# misc
-eyre.workspace = true
-url.workspace = true
-
-[dev-dependencies]
-base-builder-core = { workspace = true, features = ["test-utils"] }
-alloy-primitives.workspace = true
-tokio.workspace = true
-rstest.workspace = true
-
-[features]
-default = []
-jemalloc = [ "base-builder-core/jemalloc", "reth-optimism-cli/jemalloc" ]
