--- crates/builder/cli/src/lib.rs
@@ -1,19 +1,0 @@
-#![doc = include_str!("../README.md")]
-#![doc(issue_tracker_base_url = "https://github.com/base/base/issues/")]
-#![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
-#![cfg_attr(not(test), warn(unused_crate_dependencies))]
-
-mod flashblocks;
-pub use flashblocks::FlashblocksArgs;
-
-mod builder;
-pub use builder::BuilderArgs;
-
-mod telemetry;
-pub use telemetry::TelemetryArgs;
-
-mod wrapper;
-pub use wrapper::{Cli, RethCli};
-
-mod launcher;
-pub use launcher::BuilderLauncher;
